Support extensions in TextContentRenderer
