+++ commonmark/src/main/java/org/commonmark/content/renderer/TextContentNodeRenderer.java
@@ -1,0 +1,258 @@
+package org.commonmark.content.renderer;
+
+import org.commonmark.content.TextContentWriter;
+import org.commonmark.node.*;
+import org.commonmark.renderer.NodeRenderer;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+
+/**
+ * The node renderer that renders all the core nodes (comes last in the order of node renderers).
+ */
+public class TextContentNodeRenderer extends AbstractVisitor implements NodeRenderer {
+
+    protected final TextContentNodeRendererContext context;
+    private final TextContentWriter textContent;
+
+    private Integer orderedListCounter;
+    private Character orderedListDelimiter;
+
+    private Character bulletListMarker;
+
+    public TextContentNodeRenderer(TextContentNodeRendererContext context) {
+        this.context = context;
+        this.textContent = context.getWriter();
+    }
+
+    @Override
+    public Set<Class<? extends Node>> getNodeTypes() {
+        return new HashSet<>(Arrays.asList(
+                Document.class,
+                Heading.class,
+                Paragraph.class,
+                BlockQuote.class,
+                BulletList.class,
+                FencedCodeBlock.class,
+                HtmlBlock.class,
+                ThematicBreak.class,
+                IndentedCodeBlock.class,
+                Link.class,
+                ListItem.class,
+                OrderedList.class,
+                Image.class,
+                Emphasis.class,
+                StrongEmphasis.class,
+                Text.class,
+                Code.class,
+                HtmlInline.class,
+                SoftLineBreak.class,
+                HardLineBreak.class
+        ));
+    }
+
+    @Override
+    public void render(Node node) {
+        node.accept(this);
+    }
+
+    @Override
+    public void visit(Document document) {
+        // No rendering itself
+        visitChildren(document);
+    }
+
+    @Override
+    public void visit(BlockQuote blockQuote) {
+        textContent.write('«');
+        visitChildren(blockQuote);
+        textContent.write('»');
+
+        writeEndOfLine(blockQuote, null);
+    }
+
+    @Override
+    public void visit(BulletList bulletList) {
+        bulletListMarker = bulletList.getBulletMarker();
+        visitChildren(bulletList);
+        writeEndOfLine(bulletList, null);
+        bulletListMarker = null;
+    }
+
+    @Override
+    public void visit(Code code) {
+        textContent.write('\"');
+        textContent.write(code.getLiteral());
+        textContent.write('\"');
+    }
+
+    @Override
+    public void visit(FencedCodeBlock fencedCodeBlock) {
+        if (context.stripNewlines()) {
+            textContent.writeStripped(fencedCodeBlock.getLiteral());
+            writeEndOfLine(fencedCodeBlock, null);
+        } else {
+            textContent.write(fencedCodeBlock.getLiteral());
+        }
+    }
+
+    @Override
+    public void visit(HardLineBreak hardLineBreak) {
+        writeEndOfLine(hardLineBreak, null);
+    }
+
+    @Override
+    public void visit(Heading heading) {
+        visitChildren(heading);
+        writeEndOfLine(heading, ':');
+    }
+
+    @Override
+    public void visit(ThematicBreak thematicBreak) {
+        if (!context.stripNewlines()) {
+            textContent.write("***");
+        }
+        writeEndOfLine(thematicBreak, null);
+    }
+
+    @Override
+    public void visit(HtmlInline htmlInline) {
+        writeText(htmlInline.getLiteral());
+    }
+
+    @Override
+    public void visit(HtmlBlock htmlBlock) {
+        writeText(htmlBlock.getLiteral());
+    }
+
+    @Override
+    public void visit(Image image) {
+        writeLink(image, image.getTitle(), image.getDestination());
+    }
+
+    @Override
+    public void visit(IndentedCodeBlock indentedCodeBlock) {
+        if (context.stripNewlines()) {
+            textContent.writeStripped(indentedCodeBlock.getLiteral());
+            writeEndOfLine(indentedCodeBlock, null);
+        } else {
+            textContent.write(indentedCodeBlock.getLiteral());
+        }
+    }
+
+    @Override
+    public void visit(Link link) {
+        writeLink(link, link.getTitle(), link.getDestination());
+    }
+
+    @Override
+    public void visit(ListItem listItem) {
+        if (orderedListCounter != null) {
+            textContent.write(String.valueOf(orderedListCounter) + orderedListDelimiter + " ");
+            visitChildren(listItem);
+            writeEndOfLine(listItem, null);
+            orderedListCounter++;
+        } else if (bulletListMarker != null) {
+            if (!context.stripNewlines()) {
+                textContent.write(bulletListMarker + " ");
+            }
+            visitChildren(listItem);
+            writeEndOfLine(listItem, null);
+        }
+    }
+
+    @Override
+    public void visit(OrderedList orderedList) {
+        orderedListCounter = orderedList.getStartNumber();
+        orderedListDelimiter = orderedList.getDelimiter();
+        visitChildren(orderedList);
+        writeEndOfLine(orderedList, null);
+        orderedListCounter = null;
+        orderedListDelimiter = null;
+    }
+
+    @Override
+    public void visit(Paragraph paragraph) {
+        visitChildren(paragraph);
+        // Add "end of line" only if its "root paragraph.
+        if (paragraph.getParent() == null || paragraph.getParent() instanceof Document) {
+            writeEndOfLine(paragraph, null);
+        }
+    }
+
+    @Override
+    public void visit(SoftLineBreak softLineBreak) {
+        writeEndOfLine(softLineBreak, null);
+    }
+
+    @Override
+    public void visit(Text text) {
+        writeText(text.getLiteral());
+    }
+
+    @Override
+    protected void visitChildren(Node parent) {
+        Node node = parent.getFirstChild();
+        while (node != null) {
+            Node next = node.getNext();
+            context.render(node);
+            node = next;
+        }
+    }
+
+    private void writeText(String text) {
+        if (context.stripNewlines()) {
+            textContent.writeStripped(text);
+        } else {
+            textContent.write(text);
+        }
+    }
+
+    private void writeLink(Node node, String title, String destination) {
+        boolean hasChild = node.getFirstChild() != null;
+        boolean hasTitle = title != null;
+        boolean hasDestination = destination != null && !destination.equals("");
+
+        if (hasChild) {
+            textContent.write('"');
+            visitChildren(node);
+            textContent.write('"');
+            if (hasTitle || hasDestination) {
+                textContent.whitespace();
+                textContent.write('(');
+            }
+        }
+
+        if (hasTitle) {
+            textContent.write(title);
+            if (hasDestination) {
+                textContent.colon();
+                textContent.whitespace();
+            }
+        }
+
+        if (hasDestination) {
+            textContent.write(destination);
+        }
+
+        if (hasChild && (hasTitle || hasDestination)) {
+            textContent.write(')');
+        }
+    }
+
+    private void writeEndOfLine(Node node, Character c) {
+        if (context.stripNewlines()) {
+            if (c != null) {
+                textContent.write(c);
+            }
+            if (node.getNext() != null) {
+                textContent.whitespace();
+            }
+        } else {
+            if (node.getNext() != null) {
+                textContent.line();
+            }
+        }
+    }
+}
