+++ commonmark/src/main/java/org/commonmark/content/renderer/TextContentNodeRendererFactory.java
@@ -1,0 +1,6 @@
+package org.commonmark.content.renderer;
+
+import org.commonmark.renderer.NodeRendererFactory;
+
+public interface TextContentNodeRendererFactory extends NodeRendererFactory<TextContentNodeRendererContext> {
+}
